Rearranging Zuul modules (#305)
* Separating functions that perform the Zuul queries from the model generators.

* Moving location of zuul api interfaces.

* Grouping zuul modules into packages.

* Renaming queries.

* Renaming 'requests' module.

* Renaming 'queries' and 'models' modules.

* Adding a graph with relationship between zuul modules.
