+++ cibyl/sources/zuul/apis/utils/tests/ansible/parser.py
@@ -16,9 +16,9 @@
 import dateutil.parser
 
 from cibyl.models.ci.zuul.tests.ansible import AnsibleTestStatus
-from cibyl.sources.zuul.apis.tests.ansible.types import (AnsibleTest,
-                                                         AnsibleTestHost)
-from cibyl.sources.zuul.apis.tests.types import TestSuite
+from cibyl.sources.zuul.apis.utils.tests.ansible.types import (AnsibleTest,
+                                                               AnsibleTestHost)
+from cibyl.sources.zuul.apis.utils.tests.types import TestSuite
 from cibyl.utils.json import Draft7ValidatorFactory, JSONValidatorFactory
